Use exhibit_builder_render_exhibit_page() instead of customized version
When trying to access a page in Omeka 2.2.2 with Exhibit Builder 3.1.3
for an exhibition using cul-general, the following would show up on the
content part of the page (the lhs sidebar with the navigating links
would still show up):
This page does not have a layout.
The above is due to the fact that the customized version,
cul_legacy_exhibit_builder_render_exhibit_page(), tries
to access the layout directly, but outputs the above line
if it cannot figure out the layout for the page: in this new version
of ExhibitBuilder, the page contains content blocks, and each content block
can have a different layout -- there is no concept of an exhibit page
layout.
The call to
cul_legacy_exhibit_builder_render_exhibit_page() was replaced with
a call to exhibit_builder_render_exhibit_page(), and everything seems
to work fine.