Add CUL entry for the library unit config field
Before this commit, the library unit specified actual libraries
such as RBML, Burke, etc. However, the cul-general theme can used as
the default theme, and it may called outside of an exhibition setting.
In this case, there is no way to know what the library unit config field
was set to in the exhibit containing the item, or the item was never
included in an exhibit. Therefore, a generic cul value was added
to the library unit option in config.ini.
Historical note: this need came about when the cul-show-collection
was implemented, where an exhibition's home page will list the items
in a collection, but these items are not in the exhibition.