Change image background color on home page burgundy color scheme
Turns out the background color behind the image on the home page
for an exhibition using the burgundy color scheme was red, not
burgundy. This never caused a problem because the images took
up all the available space and the background was therefore hidden.
However, this becomes a visual problem if the background is visible
with a smaller image.